Integrate doctests with pytest
